+++ src/array/index.ts
@@ -1,2 +1,2 @@
 export * from './array.js'
-export * from './items/items.js'
+export * from './items.js'