+++ src/array/items.ts
@@ -1,8 +1,8 @@
-import { handleUnknownError } from '../../_internal/parseUnknownError/index.js'
-import type { YobtaItems } from '../../_types/YobtaItems.js'
-import type { PipeFactoryResult } from '../../_types/YobtaPipe.js'
-import type { YobtaSyncRule, YobtaSyncRules } from '../../ruleYobta/index.js'
-import { ruleYobta } from '../../ruleYobta/index.js'
+import { handleUnknownError } from '../_internal/parseUnknownError/index.js'
+import type { YobtaItems } from '../_types/YobtaItems.js'
+import type { PipeFactoryResult } from '../_types/YobtaPipe.js'
+import type { YobtaSyncRule, YobtaSyncRules } from '../ruleYobta/index.js'
+import { ruleYobta } from '../ruleYobta/index.js'
 
 export const items: YobtaItems = <R extends YobtaSyncRules>(
   ...rules: R