+++ src/requiredYobta/index.test.ts
@@ -1,10 +1,10 @@
 /* eslint-disable import/extensions */
 import { createValidator } from '../createValidator/createValidator'
-import { stringYobta } from '../stringYobta'
+import { string } from '../string'
 import { requiredMessage, requiredYobta } from './'
 
 const customMessage = 'yobta!'
-const validate = createValidator(requiredYobta(customMessage), stringYobta())
+const validate = createValidator(requiredYobta(customMessage), string())
 
 it('accepts value', () => {
   const result = validate('yobta')
@@ -17,7 +17,7 @@ it('rejects undefined', () => {
 })
 
 it('has default error message', () => {
-  const validateDefault = createValidator(requiredYobta(), stringYobta())
+  const validateDefault = createValidator(requiredYobta(), string())
   const attempt = (): any => validateDefault(undefined)
   expect(attempt).toThrow(requiredMessage)
 })