+++ src/string/index.ts
@@ -1,2 +1,3 @@
 export * from './email.js'
 export * from './reEmail.js'
+export * from './string.js'