+++ src/string/string.ts
@@ -9,7 +9,7 @@ const coercedTypes = new Set(['number', 'boolean', 'string'])
 
 export const stringMessage = 'It should be a string'
 
-export const stringYobta: StringFactory = (message = stringMessage) =>
+export const string: StringFactory = (message = stringMessage) =>
   ruleYobta<unknown, string>((value = '') => {
     if (value instanceof String || coercedTypes.has(typeof value)) {
       return String(value).trim()